+++ .dockerignore
@@ -1,2 +1,2 @@
-chat-langchain/
+frontend/
 assets/
