+++ frontend/app/components/ChatMessageBubble.tsx
