+++ frontend/app/components/InlineCitation.tsx
