+++ frontend/app/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { v4 as uuidv4 } from "uuid";
-import { ChatWindow } from "../app/components/ChatWindow";
+import { ChatWindow } from "./components/ChatWindow";
 import { ToastContainer } from "react-toastify";
 
 import { ChakraProvider } from "@chakra-ui/react";
