+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "chat-langchain",
+  "name": "frontend",
   "version": "0.1.0",
   "private": true,
   "packageManager": "yarn@1.22.19",
